refactor(widgets): ResultsPerPage -> HitsPerPage
