+++ examples/dev-novel/main.ts
@@ -207,14 +207,14 @@ storiesOf("ResultsPerPage")
     "default",
     wrapWithHits({
       template: `
-      <ng-ais-results-per-page
+      <ng-ais-hits-per-page
         [items]="[
           { value: 3, label: '3 per page' },
           { value: 5, label: '5 per page' },
           { value: 10, label: '10 per page' }
         ]"
       >
-      </ng-ais-results-per-page>
+      </ng-ais-hits-per-page>
     `
     })
   )
@@ -222,14 +222,14 @@ storiesOf("ResultsPerPage")
     "with default to 5",
     wrapWithHits({
       template: `
-        <ng-ais-results-per-page
+        <ng-ais-hits-per-page
           [items]="[
             { value: 3, label: '3 per page' },
             { value: 5, label: '5 per page', default: true },
             { value: 10, label: '10 per page' }
           ]"
         >
-        </ng-ais-results-per-page>
+        </ng-ais-hits-per-page>
       `
     })
   );
