+++ src/index.ts
@@ -5,7 +5,7 @@ import { NgAisBreadcrumbModule } from "./breadcrumb/breadcrumb.module";
 import { NgAisClearRefinementsModule } from "./clear-refinements/clear-refinements.module";
 import { NgAisCurrentRefinementsModule } from "./current-refinements/current-refinements.module";
 import { NgAisHierarchicalMenuModule } from "./hierarchical-menu/hierarchical-menu.module";
-import { NgAisHitsPerPageModule } from "./results-per-page/results-per-page.module";
+import { NgAisHitsPerPageModule } from "./hits-per-page/hits-per-page.module";
 import { NgAisHitsModule } from "./hits/hits.module";
 import { NgAisInfiniteResultsModule } from "./infinite-results/infinite-results.module";
 import { NgAisInstantSearchModule } from "./instantsearch/instantsearch.module";
